+++ src/main/java/org/stellar/sdk/Transaction.java
@@ -3,6 +3,8 @@ package org.stellar.sdk;
 import com.google.common.base.Objects;
 import org.stellar.sdk.xdr.*;
 
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -90,6 +92,42 @@ public class Transaction extends AbstractTransaction {
    */
   public Operation[] getOperations() {
     return mOperations;
+  }
+
+  /**
+   * Returns the claimable balance ID for the CreateClaimableBalanceOperation at the given index within the transaction.
+   */
+  public String getClaimableBalanceId(int index) throws IOException {
+    if (index < 0 || index >= mOperations.length) {
+      throw new IllegalArgumentException("index: " + index + " is outside the bounds of the operations within this transaction");
+    }
+    if (!(mOperations[index] instanceof CreateClaimableBalanceOperation)) {
+      throw new IllegalArgumentException("operation at index " + index + " is not of type CreateClaimableBalanceOperation: " + mOperations[index].getClass());
+    }
+
+    // We mimic the relevant code from Stellar Core
+    // https://github.com/stellar/stellar-core/blob/9f3cc04e6ec02c38974c42545a86cdc79809252b/src/test/TestAccount.cpp#L285
+    //
+    // Note that the source account must be *unmuxed* for this to work.
+
+    OperationID id = new OperationID();
+    id.setDiscriminant(EnvelopeType.ENVELOPE_TYPE_OP_ID);
+    OperationID.OperationIDId body = new OperationID.OperationIDId();
+    body.setOpNum(new Uint32(index));
+    body.setSeqNum(new SequenceNumber(new Int64(getSequenceNumber())));
+    body.setSourceAccount(AccountConverter.disableMuxed().encode(getSourceAccount()));
+    id.setId(body);
+
+    ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+    id.encode(new XdrDataOutputStream(outputStream));
+
+    ClaimableBalanceID result = new ClaimableBalanceID();
+    result.setDiscriminant(ClaimableBalanceIDType.CLAIMABLE_BALANCE_ID_TYPE_V0);
+    result.setV0(new Hash(Util.hash(outputStream.toByteArray())));
+
+    outputStream.reset();
+    result.encode(new XdrDataOutputStream(outputStream));
+    return Util.bytesToHex(outputStream.toByteArray()).toLowerCase();
   }
 
   /**