+++ CHANGELOG.md
@@ -5,6 +5,7 @@ As this project is pre 1.0, breaking changes may happen for minor version bumps.
 ## Unreleased
 
 * Add opt-in support for [SEP23](https://github.com/stellar/stellar-protocol/blob/master/ecosystem/sep-0023.md) M-strkeys for `MuxedAccount`s
+* Add `getClaimableBalanceId()` method to `Transaction` class which returns the claimable balance id for a given operation.
 
 ## 0.25.0